Upgrade more packages

+++ check.sh
@@ -3,18 +3,22 @@
 set -eux
 
 check_requirements() {
-    temp_file=$(mktemp)
-    trap 'rm -f "$temp_file"' EXIT
-    uv export --no-dev > "$temp_file"
-    if ! diff -u requirements.txt "$temp_file"; then
-        echo "requirements.txt is out of sync with pyproject.toml"
-        exit 1
+    local check_mode=$1
+    if [ "$check_mode" = "true" ]; then
+        temp_file=$(mktemp)
+        trap 'rm -f "$temp_file"' EXIT
+        uv export --no-dev > "$temp_file"
+        if ! diff -u requirements.txt "$temp_file"; then
+            echo "requirements.txt is out of sync with pyproject.toml"
+            exit 1
+        fi
+    else
+        uv export --no-dev > requirements.txt
     fi
 }
 
 run_lint() {
     local check_mode=$1
-    check_requirements
     if [ "$check_mode" = "true" ]; then
         uv run isort --check acoustid/ tests/
         uv run black --check acoustid/ tests/
@@ -62,6 +66,7 @@ done
 
 # Execute based on parsed arguments
 if [ "$run_linting" = "true" ]; then
+    check_requirements "$check_mode"
     run_lint "$check_mode"
 fi
